feat: Update token creation page with improved layout and buttons
