+++ UI/src/app/page/use-cases-page/pages/setup-token/components/token-creation-form/token-creation-form.component.scss
@@ -1,0 +1,10 @@
+@import './../../../../../../shared/styles/form.scss';
+
+.network-card {
+  cursor: pointer;
+}
+
+.network-card:hover,
+.network-card.active {
+  background-color: #f5f5f5;
+}
