+++ UI/src/app/page/use-cases-page/pages/setup-token/components/token-creation-form/token-creation-form.component.spec.ts
@@ -1,0 +1,22 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { TokenCreationFormComponent } from './token-creation-form.component';
+
+describe('TokenCreationFormComponent', () => {
+  let component: TokenCreationFormComponent;
+  let fixture: ComponentFixture<TokenCreationFormComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [TokenCreationFormComponent]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(TokenCreationFormComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
