+++ UI/src/app/page/use-cases-page/pages/setup-token/components/token-creation-form/token-creation-form.component.ts
@@ -1,0 +1,115 @@
+import { Component, EventEmitter, Output } from '@angular/core';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { INetworkDetail, NetworkChainId, Web3LoginService } from '@chainbrary/web3-login';
+import { CommonButtonText } from './../../../../../../shared/enum';
+import { ITokenCreationPayload } from './../../../../../../shared/interfaces';
+import { FormatService } from './../../../../../../shared/services/format/format.service';
+
+@Component({
+  selector: 'app-token-creation-form',
+  templateUrl: './token-creation-form.component.html',
+  styleUrl: './token-creation-form.component.scss'
+})
+export class TokenCreationFormComponent {
+  @Output() createTokenEvent: EventEmitter<ITokenCreationPayload> = new EventEmitter<ITokenCreationPayload>();
+  commonButtonText = CommonButtonText;
+  mainForm: FormGroup<ITokenCreationForm> = new FormGroup<ITokenCreationForm>({
+    name: new FormControl<string | null>(null, [Validators.required]),
+    symbol: new FormControl<string | null>(null, [Validators.required]),
+    network: new FormControl<NetworkChainId | null>(null, [Validators.required]),
+    maxSupply: new FormControl<number | null>(null, [Validators.required]),
+    decimals: new FormControl<number | null>(18, [Validators.required, Validators.min(1), Validators.max(18)]),
+    options: new FormGroup<ICheckboxOptionsForm>({
+      canBurn: new FormControl<boolean>(false),
+      canMint: new FormControl<boolean>(false),
+      canPause: new FormControl<boolean>(false)
+    })
+  });
+
+  networkAvailable: INetworkDetail[] = this.web3LoginService
+    .getNetworkDetailList()
+    .filter((network: INetworkDetail) =>
+      [NetworkChainId.ETHEREUM, NetworkChainId.BNB, NetworkChainId.AVALANCHE, NetworkChainId.POLYGON].includes(
+        network.chainId
+      )
+    );
+
+  optionsList: ITokenOptions[] = [
+    {
+      formControlName: 'canBurn',
+      title: $localize`:@@canBurn:Can Burn`,
+      description: $localize`:@@canBurnDescription:Allow the contract owner to burn tokens.`
+    },
+    {
+      formControlName: 'canMint',
+      title: $localize`:@@canMint:Can Mint`,
+      description: $localize`:@@canMintDescription:Allow the contract owner to mint tokens.`
+    },
+    {
+      formControlName: 'canPause',
+      title: $localize`:@@canPause:Can Pause`,
+      description: $localize`:@@canPauseDescription:Allow the contract owner to pause the contract.`
+    }
+  ];
+
+  constructor(
+    public formatService: FormatService,
+    private web3LoginService: Web3LoginService
+  ) {}
+
+  selectNetwork(network: NetworkChainId): void {
+    return this.mainForm.get('network')?.setValue(network);
+  }
+
+  submit(): void {
+    this.mainForm.markAllAsTouched();
+    if (this.mainForm.invalid) return;
+    else this.createTokenEvent.emit(this.formToTokenCreationPayload());
+  }
+
+  getTokenOptionControlByName(name: string): FormControl<boolean> {
+    return this.mainForm.get('options')?.get(name) as FormControl<boolean>;
+  }
+
+  increaseDecimals(): void {
+    return this.mainForm.get('decimals')?.setValue((this.mainForm.get('decimals')?.value as number) + 1);
+  }
+
+  decreaseDecimals(): void {
+    return this.mainForm.get('decimals')?.setValue((this.mainForm.get('decimals')?.value as number) - 1);
+  }
+
+  private formToTokenCreationPayload(): ITokenCreationPayload {
+    return {
+      name: this.mainForm.get('name')?.value as string,
+      symbol: this.mainForm.get('symbol')?.value as string,
+      network: this.mainForm.get('network')?.value as NetworkChainId,
+      maxSupply: this.mainForm.get('maxSupply')?.value as number,
+      decimals: this.mainForm.get('decimals')?.value as number,
+      canBurn: this.mainForm.get('options')?.get('canBurn')?.value as boolean,
+      canMint: this.mainForm.get('options')?.get('canMint')?.value as boolean,
+      canPause: this.mainForm.get('options')?.get('canPause')?.value as boolean
+    };
+  }
+}
+
+export interface ITokenCreationForm {
+  name: FormControl<string | null>;
+  symbol: FormControl<string | null>;
+  network: FormControl<NetworkChainId | null>;
+  maxSupply: FormControl<number | null>;
+  decimals: FormControl<number | null>;
+  options: FormGroup<ICheckboxOptionsForm>;
+}
+
+export interface ICheckboxOptionsForm {
+  canBurn: FormControl<boolean | null>;
+  canMint: FormControl<boolean | null>;
+  canPause: FormControl<boolean | null>;
+}
+
+interface ITokenOptions {
+  formControlName: string;
+  title: string;
+  description: string;
+}
