+++ UI/src/app/page/use-cases-page/pages/setup-token/containers/token-creation-page/token-creation-page.component.html
@@ -7,132 +7,7 @@
   <div class="d-flex justify-content-center">
     <div class="col-12">
       <div class="card border-0 my-2 shadow-sm p-3">
-        <div class="container">
-          <form [formGroup]="mainForm">
-            <div class="row my-2">
-              <h5 class="mb-3">Select Network</h5>
-              @for (item of networkAvailable; track item.chainId) {
-                <div class="col-3">
-                  <div
-                    class="card h-100 network-card"
-                    [ngClass]="{ active: mainForm.get('network')?.value === item.chainId }"
-                    (click)="selectNetwork(item.chainId)"
-                  >
-                    <div class="row mx-0">
-                      <div class="text-center mx-auto py-3">
-                        <img
-                          [src]="formatService.getIconUrl(item.chainId)"
-                          [alt]="formatService.getIconUrl(item.chainId)"
-                        />
-                      </div>
-                    </div>
-                    <div class="d-flex justify-content-center mb-2">
-                      {{ item.nativeCurrency.symbol }}
-                    </div>
-                  </div>
-                </div>
-              }
-              <mat-error
-                *ngIf="mainForm.get('network')?.hasError('required') && mainForm.get('network')?.touched === true"
-              >
-                <span i18n>Network is required</span>
-              </mat-error>
-            </div>
-            <div class="row my-2">
-              <h5 class="mb-3">Token details</h5>
-              <div class="col-12">
-                <div class="row">
-                  <div class="col-6">
-                    <mat-form-field class="w-100" appearance="outline">
-                      <input
-                        formControlName="name"
-                        matInput
-                        type="text"
-                        placeholder="Token Name"
-                        i18n-placeholder
-                        maxlength="12"
-                      />
-                      <mat-error *ngIf="mainForm.get('name')?.hasError('required')">
-                        <span i18n>Name required</span>
-                      </mat-error>
-                    </mat-form-field>
-                  </div>
-                  <div class="col-6">
-                    <mat-form-field class="w-100" appearance="outline">
-                      <input
-                        formControlName="symbol"
-                        matInput
-                        type="text"
-                        placeholder="Token Symbol"
-                        i18n-placeholder
-                        maxlength="5"
-                      />
-                      <mat-error *ngIf="mainForm.get('symbol')?.hasError('required')">
-                        <span i18n>Symbol is required</span>
-                      </mat-error>
-                    </mat-form-field>
-                  </div>
-                  <div class="col-12">
-                    <mat-form-field class="w-100" appearance="outline">
-                      <input
-                        formControlName="maxSupply"
-                        matInput
-                        type="number"
-                        placeholder="Max Supply"
-                        i18n-placeholder
-                      />
-                      <mat-error *ngIf="mainForm.get('maxSupply')?.hasError('required')">
-                        <span i18n>Max supply invalid</span>
-                      </mat-error>
-                    </mat-form-field>
-                  </div>
-                  <div class="col-5">
-                    <mat-form-field class="w-100" appearance="outline">
-                      <mat-label>Decimal (1-18)</mat-label>
-                      <input formControlName="decimals" class="text-center" matInput type="number" />
-                      <button matPrefix mat-icon-button type="button" (click)="decreaseDecimals()">
-                        <i class="bi bi-dash-circle"></i>
-                      </button>
-                      <button matSuffix mat-icon-button type="button" (click)="increaseDecimals()">
-                        <i class="bi bi-plus-circle"></i>
-                      </button>
-                    </mat-form-field>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div class="row my-2">
-              <h5 class="mb-3">Other Configs</h5>
-              <div class="row mb-3">
-                @for (item of optionsList; track item) {
-                  <div class="col-6">
-                    <mat-checkbox [formControl]="getTokenOptionControlByName(item.formControlName)" color="primary">
-                      {{ item.title }}
-                      <i
-                        [matTooltip]="item.description"
-                        aria-label="Button that displays a tooltip when focused or hovered over"
-                        class="ps-2 bi bi-info-circle"
-                      >
-                      </i>
-                    </mat-checkbox>
-                  </div>
-                }
-              </div>
-            </div>
-            <div class="d-flex justify-content-end my-2">
-              <div class="col-6 col-lg-4 px-2">
-                <a routerLink="./../../setup-token/services">
-                  <button class="btn btn-outline-primary w-100">
-                    {{ commonButtonText.Back }}
-                  </button>
-                </a>
-              </div>
-              <div class="col-6 col-lg-4 px-2">
-                <button i18n class="btn btn-primary w-100" (click)="submit()">Create Token</button>
-              </div>
-            </div>
-          </form>
-        </div>
+        <app-token-creation-form (createTokenEvent)="submit($event)"></app-token-creation-form>
       </div>
     </div>
   </div>
