--- UI/src/app/page/use-cases-page/pages/setup-token/containers/token-creation-page/token-creation-page.component.scss
@@ -1,10 +1,0 @@
-@import './../../../../../../shared/styles/form.scss';
-
-.network-card {
-  cursor: pointer;
-}
-
-.network-card:hover,
-.network-card.active {
-  background-color: #f5f5f5;
-}
